Pre-Removing Node SDK and CCG
The Node SDK is deprecated and the Typescript SDK should be used instead.

The Client Credentials Grant is the wrong type of authorization type for the application. Because Box developer resources are sub-par, it was not made obvious that the service worker account - which has admin privileges - cannot access folders and files made by users because it exists in its own file tree.